requirements_new.txt
# duckduckgo-search
duckduckgo-search==6.3.5

# openai (for Whisper voice transcription)
openai==1.54.0

# google-generativeai (for Gemini)
google-generativeai==0.8.3
